git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/brauer@7874 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/DATASTRUCTURES/GridFeature.C
@@ -4,7 +4,7 @@
 // --------------------------------------------------------------------------
 //                   OpenMS Mass Spectrometry Framework
 // --------------------------------------------------------------------------
-//  Copyright (C) 2003-2010 -- Oliver Kohlbacher, Knut Reinert
+//  Copyright (C) 2003-2011 -- Oliver Kohlbacher, Knut Reinert
 //
 //  This library is free software; you can redistribute it and/or
 //  modify it under the terms of the GNU Lesser General Public
@@ -58,7 +58,7 @@ namespace OpenMS
 
 	Int GridFeature::getID()
 	{
-		return feature_index_;
+		return (Int)feature_index_;
 	}
 
 }
